Examples from Component READMEs
This loader is the same one as the one in the playground.

This is very rough and ready, the tsconfig and webpack can probably be
tidied up a bit and the build can be made faster - it waits around 70%
for quite a long time for a reason I don't yet understand but it does
work eventually.