Define Props interfaces in AppProvider, Toast and ContextualSavebar
The styleguide's Props Explorer data generation script expects the props
for a component to live in an interface called Props. This meant that
these three broke the generation.

By defining a bare interface that extends the interface we actually want
we can keep the Props Explorer data generation script happy.